requirements/base.txt
django-app-settings==0.2.2
django-braces==1.11.0
